+++ src/types/Globals.ts
@@ -4,7 +4,7 @@ import { RaceVars } from "../features/race/types/RaceVars";
 import { ChatMessage } from "./ChatMessage";
 
 export class Globals {
-  debug = false;
+  debug = true;
 
   // Cached API functions
   g = game;
@@ -12,7 +12,7 @@ export class Globals {
   r = game.GetRoom();
   seeds = game.GetSeeds();
   itemPool = game.GetItemPool();
-  fonts = {
+  readonly fonts = {
     droid: Font(),
     pf: Font(),
   };