+++ scripts/get_version_from_package_json.py
@@ -1,6 +1,7 @@
 import json
 import os
-import sys
+
+from utils import error
 
 SCRIPT_PATH = os.path.realpath(__file__)
 SCRIPT_DIRECTORY = os.path.dirname(SCRIPT_PATH)
@@ -15,12 +16,3 @@ def get_version_from_package_json():
         error('Failed to find the version in the "{}" file.'.format(PACKAGE_JSON_PATH))
 
     return package_json["version"]
-
-
-def error(msg):
-    printf("Error: {}".format(msg))
-    sys.exit(1)
-
-
-def printf(*args):
-    print(*args, flush=True)